Refactor: move fileGroupPathToIdentifier to tools, fix component types
Agent-Logs-Url: https://github.com/Vempain/vempain-file-frontend/sessions/4e73dfaa-e5ba-4f9b-806c-d46f8b2d2ed8
