+++ src/components/data/index.ts
@@ -1,2 +1,2 @@
 export {PublishMusicData} from "./PublishMusicData";
-export {PublishGpsTimeSeries, fileGroupPathToIdentifier} from "./PublishGpsTimeSeries";
+export {PublishGpsTimeSeries} from "./PublishGpsTimeSeries";
