+++ src/tools/index.ts
@@ -1,4 +1,5 @@
 export {compareDayjsNullable, formatDayjsNullable, formatDateWithTimeZone} from './dateTimeTool';
+export {fileGroupPathToIdentifier} from './dataTools';
 export {fileTypeEnum2Tag, pathCompletionEnum2Tag} from './enumToTagTool';
 export {LanguageTool} from "./languageTool";
 export {getPaginationConfig} from './tablePaginationConfig';
